requirements.txt
pandas
requests
mysql-connector-python
python-dotenv
datetime
apache-airflow